refactored: UserProfile is now a toggle component that switches between view/edit components
